Projected list feature

+++ src/main/java/org/xmlbeam/AutoMap.java
@@ -41,6 +41,7 @@ import org.xmlbeam.types.XBAutoMap;
 import org.xmlbeam.util.intern.DOMHelper;
 import org.xmlbeam.util.intern.duplex.DuplexExpression;
 import org.xmlbeam.util.intern.duplex.DuplexXPathParser;
+import org.xmlbeam.util.intern.duplex.ExpressionType;
 
 /**
  * @author sven
@@ -61,7 +62,11 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
     private final DomChangeTracker domChangeTracker = new DomChangeTracker() {
         @Override
         void refresh(final boolean forWrite) throws XPathExpressionException {
-            final NodeList nodes = (NodeList) invocationContext.getxPathExpression().evaluate(baseNode, XPathConstants.NODESET);;
+            if (invocationContext.getxPathExpression() == null) {
+                boundNode = baseNode;
+                return;//xPath expression is optional for maps
+            }
+            final NodeList nodes = (NodeList) invocationContext.getxPathExpression().evaluate(baseNode, XPathConstants.NODESET);
             if ((nodes.getLength() == 0) && forWrite) {
                 boundNode = invocationContext.getDuplexExpression().ensureExistence(baseNode);
             } else {
@@ -70,15 +75,31 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
         }
     };
 
+    private final Class<?> valueType;
+
     /**
      * @param baseNode
      * @param invocationContext
+     * @param valueType
+     *            map value type
      */
-    public AutoMap(final Node baseNode, final InvocationContext invocationContext) {
+    public AutoMap(final Node baseNode, final InvocationContext invocationContext, final Class<?> valueType) {
         this.invocationContext = invocationContext;
         this.baseNode = baseNode;
         this.invocationContext.getProjector().addDOMChangeListener(this);
         this.typeConverter = invocationContext.getProjector().config().getTypeConverter();
+        this.valueType = valueType;
+    }
+
+    /**
+     * @see java.util.AbstractMap#clear()
+     */
+    @Override
+    public void clear() {
+        domChangeTracker.refreshForReadIfNeeded();
+        if (boundNode != null) {
+            DOMHelper.removeAllChildren(boundNode);
+        }
     }
 
     /**
@@ -104,22 +125,73 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
      *            relative xpath
      * @return value in DOM tree. null if no value is present.
      */
+    @SuppressWarnings("unchecked")
+    @Override
     public T get(final CharSequence path) {
+        return (T) get(path, invocationContext.getTargetComponentType());
+    }
+
+    @Override
+    public <E> E get(final CharSequence path, final Class<E> asType) {
         if ((path == null) || (path.length() == 0)) {
             throw new IllegalArgumentException("Parameter path must not be empty or null");
         }
         domChangeTracker.refreshForReadIfNeeded();
+        if (boundNode == null) {
+            // we need not to evaluate, because our context node does not even exist.
+            return null;
+        }
 
         final Document document = DOMHelper.getOwnerDocumentFor(baseNode);
         final DuplexExpression duplexExpression = new DuplexXPathParser(invocationContext.getProjector().config().getUserDefinedNamespaceMapping()).compile(path);
         try {
             final XPathExpression expression = invocationContext.getProjector().config().createXPath(document).compile(duplexExpression.getExpressionAsStringWithoutFormatPatterns());
             Node prevNode = (Node) expression.evaluate(boundNode, XPathConstants.NODE);
-            final T value = DefaultXPathEvaluator.convertToComponentType(invocationContext, prevNode, invocationContext.getTargetComponentType());
+            InvocationContext tempContext = new InvocationContext(invocationContext.getResolvedXPath(), invocationContext.getxPath(), expression, duplexExpression, null, asType, invocationContext.getProjector());
+            final E value = DefaultXPathEvaluator.convertToComponentType(tempContext, prevNode, asType);
             return value;
         } catch (XPathExpressionException e) {
             throw new XBPathException(e, path);
         }
+    }
+
+    /**
+     * @deprecated
+     * @see java.util.AbstractMap#containsKey(java.lang.Object)
+     */
+    @Deprecated
+    @Override
+    public boolean containsKey(final Object path) {
+        if (!(path instanceof CharSequence)) {
+            throw new IllegalArgumentException("parameter path must be a CharSequence containing a relative XPath expression.");
+        }
+        return containsKey(CharSequence.class.cast(path));
+    }
+
+    /**
+     * Checks existence of value at given xpath.
+     *
+     * @param path
+     * @return true if nonnull value exists at given path
+     */
+    @Override
+    public boolean containsKey(final CharSequence path) {
+        return get(path) != null;
+    }
+
+    /**
+     * Like map.containsValue, but this map can not store null values.
+     *
+     * @param value
+     * @return true if value is found in any element or attribute
+     * @see java.util.AbstractMap#containsValue(java.lang.Object)
+     */
+    @Override
+    public boolean containsValue(final Object value) {
+        if (value == null) {
+            return false;
+        }
+        return super.containsValue(value);
     }
 
     /**
@@ -133,9 +205,18 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
         if (value == null) {
             return remove(path);
         }
+        if (boundNode == null) {
+            // If there is no context node yet, ignore if
+            // we had read the dom before. We need to create it now.
+            domChangeTracker.domChanged();
+        }
         domChangeTracker.refreshForWriteIfNeeded();
+        assert boundNode != null : "Bound node does not exist. No evaluation possible";
         final Document document = DOMHelper.getOwnerDocumentFor(baseNode);
         final DuplexExpression duplexExpression = new DuplexXPathParser(invocationContext.getProjector().config().getUserDefinedNamespaceMapping()).compile(path);
+        if ((ExpressionType.ATTRIBUTE == duplexExpression.getExpressionType()) && ProjectionInvocationHandler.isStructureChangingType(valueType)) {
+            throw new IllegalArgumentException("Value of type " + valueType + "can not be written to XML attributes. Choose a different xpath expression or use a different map component type");
+        }
         try {
             final XPathExpression expression = invocationContext.getProjector().config().createXPath(document).compile(duplexExpression.getExpressionAsStringWithoutFormatPatterns());
             Node prevNode = (Node) expression.evaluate(boundNode, XPathConstants.NODE);
@@ -153,11 +234,13 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
                     if (!(value instanceof DOMAccess)) {
                         throw new IllegalArgumentException("Parameter value is not a subprojection.");
                     }
-                    parent.appendChild(DOMAccess.class.cast(value).getDOMNode());
+                    // Dont't add the value, add a copy.
+                    //parent.appendChild(DOMAccess.class.cast(value).getDOMNode());
+                    DOMHelper.appendClone(parent, DOMAccess.class.cast(value).getDOMNode());
                 }
                 return previousValue;
             }
-
+            assert boundNode != null : "Without bound node, there is no context where something could be created in.";
             Node node = duplexExpression.ensureExistence(boundNode);
             node.setTextContent(value.toString());
 
@@ -177,7 +260,7 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
     @Override
     public T remove(final Object path) {
         if (!(path instanceof CharSequence)) {
-            throw new IllegalArgumentException("parameter path must be a CharSequence containing a relative XPath expression.");
+            throw new IllegalArgumentException("parameter path must be a CharSequence or String containing a relative XPath expression.");
         }
         return remove(CharSequence.class.cast(path));
     }
@@ -199,6 +282,9 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
         try {
             final XPathExpression expression = invocationContext.getProjector().config().createXPath(document).compile(duplexExpression.getExpressionAsStringWithoutFormatPatterns());
             Node prevNode = (Node) expression.evaluate(boundNode, XPathConstants.NODE);
+            if (prevNode == null) {
+                return null;
+            }
             final T value = DefaultXPathEvaluator.convertToComponentType(invocationContext, prevNode, invocationContext.getTargetComponentType());
             duplexExpression.deleteAllMatchingChildren(prevNode.getParentNode());
             return value;
@@ -225,7 +311,9 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
         if (boundNode == null) {
             return Collections.emptySet();
         }
-        final Set<Map.Entry<String, T>> set = new TreeSet<Map.Entry<String, T>>(ENTRY_COMPARATOR);
+        final Set<Map.Entry<String, T>> set = new TreeSet<Map.Entry<String, T>>(ENTRY_COMPARATOR) {
+
+        };
         if ((invocationContext.getTargetComponentType().isInterface()) || (Node.class.isAssignableFrom(invocationContext.getTargetComponentType()))) {
             collectChildren(set, boundNode, ".");
         } else {
@@ -234,7 +322,7 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
         return set;
     }
 
-    private <T> void collectChildren(final Set<Entry<String, T>> set, final Node n, final String path) {
+    private <E> void collectChildren(final Set<Entry<String, E>> set, final Node n, final String path) {
         if (n.getNodeType() != Node.ELEMENT_NODE) {
             return;
         }
@@ -248,15 +336,15 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
                 continue;
             }
             String childPath = path + "/" + child.getNodeName();
-            T value = DefaultXPathEvaluator.convertToComponentType(invocationContext, child, invocationContext.getTargetComponentType());
+            E value = DefaultXPathEvaluator.convertToComponentType(invocationContext, child, invocationContext.getTargetComponentType());
             if (value != null) {
-                set.add(new SimpleEntry<String, T>(childPath, value));
+                set.add(new SimpleEntry<String, E>(childPath, value));
             }
             collectChildren(set, child, childPath);
         }
     }
 
-    private <T> void collectChildrenValues(final Set<Entry<String, T>> set, final Node n, final String path) {
+    private <E> void collectChildrenValues(final Set<Entry<String, E>> set, final Node n, final String path) {
         if (n.getNodeType() == Node.TEXT_NODE) {
             return;
         }
@@ -266,7 +354,7 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
             if (attributes != null) {
                 for (int i = 0; i < attributes.getLength(); ++i) {
                     // map.put(path + "/@" + attributes.item(i).getNodeName(), attributes.item(i));
-                    set.add(new SimpleEntry<String, T>(path + "/@" + attributes.item(i).getNodeName(), (T) (attributes.item(i).getNodeValue())));
+                    set.add(new SimpleEntry<String, E>(path + "/@" + attributes.item(i).getNodeName(), (E) (attributes.item(i).getNodeValue())));
                 }
             }
         }
@@ -283,15 +371,23 @@ public class AutoMap<T> extends AbstractMap<String, T> implements XBAutoMap<T>, 
             if (typeConverter.isConvertable(invocationContext.getTargetComponentType())) {
                 final String stringContent = DOMHelper.directTextContent(child);
                 if ((stringContent != null) && (!stringContent.isEmpty())) {
-                    final T value = (T) typeConverter.convertTo(invocationContext.getTargetComponentType(), stringContent, invocationContext.getExpressionFormatPattern());
+                    final E value = (E) typeConverter.convertTo(invocationContext.getTargetComponentType(), stringContent, invocationContext.getExpressionFormatPattern());
                     //T value = DefaultXPathEvaluator.convertToComponentType(invocationContext, child, invocationContext.getTargetComponentType());
                     if (value != null) {
-                        set.add(new SimpleEntry<String, T>(childPath, value));
+                        set.add(new SimpleEntry<String, E>(childPath, value));
                     }
                 }
             }
             collectChildrenValues(set, child, childPath);
         }
+    }
+
+    /**
+     * @return bound node for this automap
+     */
+    public Node getNode() {
+        domChangeTracker.refreshForReadIfNeeded();
+        return boundNode;
     }
 
 }
